requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==4.2.0
bitarray==2.7.3
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
circlify==0.15.0
ckdl==0.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.2
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.6.0
customtkinter==5.1.3
Cython==0.29.35
daff==1.3.46
darkdetect==0.8.0
datasets==2.12.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
Django==4.1.7
dnspython==2.6.1
docker-pycreds==0.4.0
docx-mailmerge==0.5.0
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
GeoAlchemy2==0.14.4
geojson==3.0.1
geomdl==5.3.1
gitdb==4.0.10
GitPython==3.1.31
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
ifcopenshell==0.7.0.230523
imageio==2.29.0
img2pdf==0.4.4
infinity==1.5
intervals==0.9.2
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonfield==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kombu==5.3.4
lazy_loader==0.2
llvmlite==0.40.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
NET==2.4
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.57.0
numpy==1.24.2
openpyxl==3.1.0
ordered-set==4.1.0
outcome==1.2.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
piexif==1.1.3
pikepdf==8.4.0
Pillow==9.4.0
pillow-heif==0.13.1
platformdirs==3.0.0
plumbum==1.8.2
ply==3.11
premailer==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.23.2
psutil==5.9.4
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
PyJWT==2.8.0
PyLaTeX==1.4.1
PyMCubes==0.1.4
pynput==1.7.6
pyOpenSSL==23.3.0
pyOpenSubdiv==0.0.3
PyP6Xer==1.14.3
PyPDF2==3.0.1
pypng==0.20220715.0
pyproj==3.6.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.5
PythonVideoConverter==1.0.3
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
redis==5.0.1
regex==2023.5.5
requests==2.29.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.20.0
scipy==1.10.1
selenium==4.11.2
Send2Trash==1.8.0
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==3.7.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.26
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.12
tablepyxl==0.6.1
tabulate==0.9.0
terminado==0.17.1
tifffile==2023.4.12
tiktoken==0.4.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
tomlkit==0.12.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.15
validators==0.28.1
vine==5.1.0
wandb==0.15.3
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==4.0.0
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.5
wsproto==1.2.0
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
xer-reader==0.1.5
xmljson==0.2.1
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
